samurai-examples/testcase/testcase/www/html/testcases/css3/flexbox/align-absolute-child-expected.txt

PASS beforePosition[key] is afterPosition[key]
PASS beforePosition[key] is afterPosition[key]
PASS beforePosition[key] is afterPosition[key]
PASS beforePosition[key] is afterPosition[key]
PASS beforePosition[key] is afterPosition[key]
PASS beforePosition[key] is afterPosition[key]
PASS successfullyParsed is true

TEST COMPLETE
PASS
PASS
PASS
PASS
PASS
PASS
PASS
PASS
PASS
